document_semantic_matching/requirements.txt
asyncpg==0.30.0
fastapi==0.121.1
numpy==2.3.4
pandas==2.3.3
pgvector==0.4.0
protobuf==6.33.0
psycopg2==2.9.10
pydantic==2.12.4
pydantic_settings==2.11.0
pytest==8.4.2
python-dotenv==1.2.1
uvicorn==0.38.0
google-genai==1.49.0

